QUERY CREATE AND UPDATE, IMPORTANT COMMIT.
I included detection limit, standard, database query, modal for deletion, update method to update samples in plusmanager, keep variables in view, add new fields in presamples for control and timestams, schedule configuration in America Santiago, update of the migration of presamples, updates when modifying aliquot, colorimetry and absorbance. evaluation in the table if the sample is STD shows what it is. QUERY CREATE AND UPDATE, IMPORTANT COMMIT.